fix: AVP code and vendor id are unsigned ints

+++ sparta-diameter-3gpp-sgdgdd/src/test/java/com/sipgate/sparta/diameter/_3gpp/sgdgdd/SgdGddAVPProviderTest.java
@@ -28,7 +28,7 @@ class SgdGddAVPProviderTest {
             if (!Modifier.isStatic(field.getModifiers()) || !Modifier.isFinal(field.getModifiers())) {
                 continue;
             }
-            final int code = field.getInt(null);
+            final long code = field.getLong(null);
             if (!registeredCodes.contains(code)) {
                 missing.add(field.getName() + " (code=" + code + ")");
             }
